+++ CHANGES.rst
@@ -49,6 +49,10 @@ https://github.com/atviriduomenys/katalogas/issues/2227
 - Add custom error pages for 400, 403, 404, 500.
 - Fix django translations file.
 
+<No ticket>
+
+- Update `has_perm` to allow update organization for super users.
+
 v 1.11.3 (2026-01-05)
 ==================